+++ java/byps/src-ser-bin/com/wilutions/byps/BSerializer_13.java
@@ -46,7 +46,7 @@ public class BSerializer_13 extends BSerializer {
 	}
 
 	@SuppressWarnings({"unchecked", "rawtypes"})
-	public Object read(final Object obj1, final BInput bin1, final int version) throws BException {
+	public Object read(final Object obj1, final BInput bin1, final long version) throws BException {
 		final BInputBin bin = ((BInputBin)bin1);
 		// lengths
 		final int n0 = bin.bbuf.getLength();
@@ -67,7 +67,7 @@ public class BSerializer_13 extends BSerializer {
 	
 	@SuppressWarnings("rawtypes")
 	@Override
-	public void write(final Object obj1, final BOutput bout1, final int version) throws BException {
+	public void write(final Object obj1, final BOutput bout1, final long version) throws BException {
 		final BOutputBin bout = ((BOutputBin)bout1);
 		final java.util.Map arr = (java.util.Map) obj1;
