+++ java/byps/src-ser-bin/com/wilutions/byps/BSerializer_15.java
@@ -18,7 +18,7 @@ public class BSerializer_15 extends BSerializer {
 	}
 
 	@Override
-	public void write(final Object obj, final BOutput bout1, final int version) throws BException {
+	public void write(final Object obj, final BOutput bout1, final long version) throws BException {
 		final BOutputBin bout = ((BOutputBin)bout1);
 		InputStream is = (InputStream)obj;
 		BStreamRequest streamRequest = bout.createStreamRequest(is);
@@ -26,7 +26,7 @@ public class BSerializer_15 extends BSerializer {
 	}
 
 	@Override
-	public Object read(final Object obj1, final BInput bin1, final int version) throws BException {
+	public Object read(final Object obj1, final BInput bin1, final long version) throws BException {
 		BInputBin bin = ((BInputBin)bin1);
 		long streamId = bin.bbuf.getLong();
 		try {
