+++ java/byps/src-ser-bin/com/wilutions/byps/BSerializer_16.java
@@ -17,7 +17,7 @@ public class BSerializer_16 extends BSerializer {
 	}
 
 	@Override
-	public void write(final Object obj, final BOutput bout1, final int version) throws BException {
+	public void write(final Object obj, final BOutput bout1, final long version) throws BException {
 		final BOutputBin bout = ((BOutputBin)bout1);
 		final BRemote remote = (BRemote)obj;
 		final BTargetId targetId = remote.BRemote_getTargetId();
@@ -25,7 +25,7 @@ public class BSerializer_16 extends BSerializer {
 	}
 
 	@Override
-	public Object read(final Object obj1, final BInput bin1, final int version) throws BException {
+	public Object read(final Object obj1, final BInput bin1, final long version) throws BException {
 		BInputBin bin = ((BInputBin)bin1);
 		BRemote remote = null;
 		final BTargetId targetId = BTargetId.read(bin.bbuf.getBuffer());
